[refacto] triggers: move temporal workflows to endpoint
