+++ front/lib/triggers/temporal/webhook/workflows.ts
@@ -2,27 +2,9 @@ import { proxyActivities } from "@temporalio/workflow";
 
 import type * as activities from "./activities";
 
-const { runTriggerWebhookActivity, webhookCleanupActivity } = proxyActivities<
-  typeof activities
->({
+const { webhookCleanupActivity } = proxyActivities<typeof activities>({
   startToCloseTimeout: "5 minutes",
-  retry: {
-    nonRetryableErrorTypes: ["TriggerNonRetryableError"],
-  },
 });
-
-export async function agentTriggerWebhookWorkflow({
-  workspaceId,
-  webhookRequestId,
-}: {
-  workspaceId: string;
-  webhookRequestId: number;
-}) {
-  await runTriggerWebhookActivity({
-    workspaceId,
-    webhookRequestId,
-  });
-}
 
 export async function webhookCleanupWorkflow() {
   await webhookCleanupActivity();
